refactor(perf-platform): rename collector classes, drop multi-arch
Pre-content cleanup of /apps/perf-* before the workshop module
rewrite. The shape of both apps stays the same; this commit only
renames classes for clarity and drops the multi-architecture
plumbing that adds noise without value at workshop scale.

Collector

- AsyncProfilerAttach -> Profiler. The class is the perf-collector's
  service-shaped profiling component, not an action class. "Attach"
  is one method on it.
- DumpController -> CollectorController. Named after the collector's
  external API surface (POST /dump and any future endpoints), not
  one of the verbs it serves.
- DumpService -> CollectorService. Matches the controller.
- Drop ArchPostProcessor and META-INF/spring.factories. They existed
  only to pick /opt/perf-collector/{amd64,arm64}/ at runtime. The
  workshop ships a single-arch (linux/amd64) image - multi-arch is
  out of scope. Saves ~40 lines of code plus a Spring 4-deprecated
  EnvironmentPostProcessor.
- pom.xml drops 4 download-maven-plugin executions (arm64 fetches),
  2 maven-resources-plugin executions (arm64 staging), and one of
  two jib platforms. ~70 lines.
- application.yaml flattens /opt/perf-collector/${arch}/... to
  /opt/perf-collector/...

Analyzer

- Extract GitHubSourceCodeTool from inside AiService into its own
  top-level file. The tool is the workshop's source-citation wow
  moment in chapter 3 - it deserves a file we can point at and
  walk through. AiService keeps the system prompt and the prompt
  builder; the source-code Tool implementation is now a peer of
  PyroscopeTool.

Verified end-to-end on the live cluster: collector image rebuilt
and pushed (linux/amd64 only), DaemonSet restarted, async-profiler
attaches with the new logger names, JFR push to Pyroscope continues.
Analyzer image rebuilt and pushed, deployment restarted, on-demand
analysis runs all four lanes (CPU top, wall top, JFR, thread dump),
GitHubSourceCodeTool cites unicorn-store-spring source files in
the AI report.
