--- apps/perf-collector/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,0 @@
-org.springframework.boot.env.EnvironmentPostProcessor=\
-com.example.perf.collector.ArchPostProcessor
